# backend/api/article/spec/openapi.yaml
openapi: 3.0.0
info:
  title: article api
  description: |
    tamaco-blog 記事管理システムのバックエンドAPIサービス。
    ブログ記事のCRUD操作、カテゴリ・タグの管理、
    記事関連の各種操作を提供するRESTful APIです。
  version: 1.0.0
servers:
  - url: http://localhost:{port}/article/
    description: dev
    variables:
      port:
        default: '8080'
  - url: https://api.<stg-domain-name>/article/
    description: stg
  - url: https://api.<prd-domain-name>/article/
    description: prd
security:
  - bearerAuth: []
paths:
...
  /v1/tags:
    get:
      tags:
        - tags
      summary: タグ一覧取得
      description: すべてのタグの一覧を取得します。
      operationId: getTags
      responses:
        '200':
          description: タグ一覧の取得に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagList'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
      security: []
    post:
      tags:
        - tags
      summary: タグ作成
      description: 新しいタグを作成します。認証が必要です。
      operationId: createTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagCreate'
      responses:
        '201':
          description: タグの作成に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          description: リクエストパラメータが不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: 認証が必要
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '409':
          description: 同じIDのタグが既に存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlreadyExists'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Tag:
      title: Tag
      type: object
      description: 記事タグのデータモデル
      required:
        - id
        - name
        - slug
      properties:
        id:
          type: string
          format: uuid
          description: タグの一意識別子
          example: 456e7890-e89b-12d3-a456-426614174000
        name:
          type: string
          description: タグ名
          minLength: 1
          maxLength: 50
          example: Go
        slug:
          type: string
          description: URL用の識別子
          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
          example: golang
        usage_count:
          type: integer
          description: タグが使用されている記事数
          minimum: 0
          example: 15
        created_at:
          type: string
          format: date-time
          description: 作成日時
          example: '2024-01-01T09:00:00Z'
        updated_at:
          type: string
          format: date-time
          description: 更新日時
          example: '2024-01-02T10:30:00Z'
...
    BadRequest:
      title: BadRequest
      type: object
      description: リクエストが不正な場合のエラーレスポンス
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: エラーコード
          example: BAD_REQUEST
        message:
          type: string
          description: エラーメッセージ
          example: リクエストパラメータが不正です
        details:
          type: array
          description: エラーの詳細情報
          items:
            type: object
            properties:
              field:
                type: string
                description: エラーが発生したフィールド名
              reason:
                type: string
                description: エラーの理由
      example:
        code: BAD_REQUEST
        message: リクエストパラメータが不正です
        details:
          - field: title
            reason: タイトルは必須です
          - field: content
            reason: コンテンツは必須です
    InternalServerError:
      title: InternalServerError
      type: object
      description: サーバー内部エラーが発生した場合のエラーレスポンス
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: エラーコード
          example: INTERNAL_SERVER_ERROR
        message:
          type: string
          description: エラーメッセージ
          example: サーバー内部でエラーが発生しました
        request_id:
          type: string
          description: リクエストID（トレーシング用）
          example: req-123456789
      example:
        code: INTERNAL_SERVER_ERROR
        message: サーバー内部でエラーが発生しました
        request_id: req-123456789
...
    Unauthorized:
      title: Unauthorized
      type: object
      description: 認証が必要な場合のエラーレスポンス
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: エラーコード
          example: UNAUTHORIZED
        message:
          type: string
          description: エラーメッセージ
          example: 認証が必要です
      example:
        code: UNAUTHORIZED
        message: 認証が必要です
    AlreadyExists:
      title: AlreadyExists
      type: object
      description: リソースが既に存在する場合のエラーレスポンス
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: エラーコード
          example: ALREADY_EXISTS
        message:
          type: string
          description: エラーメッセージ
          example: 指定されたリソースは既に存在します
        resource:
          type: string
          description: 既に存在するリソースの種類
          example: article
        identifier:
          type: string
          description: 重複したリソースの識別子
          example: how-to-use-golang
      example:
        code: ALREADY_EXISTS
        message: 指定されたリソースは既に存在します
        resource: article
        identifier: how-to-use-golang
...
    TagList:
      title: TagListResponse
      type: object
      description: タグ一覧レスポンス
      required:
        - tags
      properties:
        tags:
          type: array
          description: タグのリスト
          items:
            $ref: '#/components/schemas/Tag'
      example:
        tags:
          - id: 456e7890-e89b-12d3-a456-426614174000
            name: Go
            slug: golang
            usage_count: 15
            created_at: '2024-01-01T09:00:00Z'
            updated_at: '2024-01-02T10:30:00Z'
          - id: 567e8901-e89b-12d3-a456-426614174000
            name: Docker
            slug: docker
            usage_count: 12
            created_at: '2024-01-01T09:00:00Z'
            updated_at: '2024-01-02T10:30:00Z'
          - id: 678e9012-e89b-12d3-a456-426614174000
            name: Kubernetes
            slug: kubernetes
            usage_count: 8
            created_at: '2024-01-01T09:00:00Z'
            updated_at: '2024-01-02T10:30:00Z'
    TagCreate:
      title: TagCreateRequest
      type: object
      description: タグ作成リクエスト
      required:
        - name
        - slug
      properties:
        name:
          type: string
          description: タグ名
          minLength: 1
          maxLength: 50
          example: Go
        slug:
          type: string
          description: URL用の識別子
          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
          example: golang
...